Split up more files, getting ready to add ChromaDB vector storage system
